pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bartholomew"
version = "0.0.1"
description = "Bartholomew AI - Proactive assistant kernel"
authors = [
    {name = "Taylor Paul", email = "tpaul733@gmail.com"}
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}
dependencies = [
    "fastapi>=0.104",
    "uvicorn[standard]>=0.24",
    "pydantic>=2.6",
    "python-dateutil>=2.8",
    "tzdata>=2023.3",
    "sqlalchemy>=2.0",
    "aiosqlite>=0.19",
    "tenacity>=8.2",
    "pyyaml>=6.0.1",
    "prometheus_client>=0.20.0",
    "keyring>=24.0",
]

[project.scripts]
bartholomew = "bartholomew.cli:main"
bartholomew-backfill-fts = "scripts.backfill_fts:main"

[tool.uvicorn]
factory = false

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
    "N",   # pep8-naming
    "COM", # flake8-commas
    "C4",  # flake8-comprehensions
    "ISC", # flake8-implicit-str-concat
    "PL",  # pylint
]
ignore = [
    "E501",   # line too long (handled by black)
    "PLR0913", # too many arguments
    "PLR2004", # magic value comparison
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # Allow unused imports in __init__.py
"identity_interpreter/models.py" = ["N801"]  # Allow Identity_ for disambiguation

[tool.ruff.lint.isort]
known-first-party = ["identity_interpreter"]
force-single-line = false
lines-after-imports = 2

[tool.pytest.ini_options]
testpaths = ["tests", "."]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-q -m 'not integration and not slow'"
markers = [
    "integration: long/slow external-ish tests",
    "slow: expensive tests",
    "smoke: fast sanity checks",
    "asyncio: mark test as async",
]
